applied gate for authorised access

+++ app/Http/Controllers/QuestionController.php
@@ -72,6 +72,9 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
+        if(\Gate::denies('update-question', $question)){
+            abort(403, 'Access Denied!');
+        }
         return view('question.edit', compact('question'));
     }
 
@@ -84,6 +87,9 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question)){
+            abort(403, 'Access Denied!');
+        }
         $question->update($request->only('title','body'));
         return redirect()->route('questions.index')->with('success','Question updated successfully');
     }
@@ -96,6 +102,9 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question', $question)){
+            abort(403, 'Access Denied!');
+        }
         $question->delete();
         return redirect()->route('questions.index')->with('success','question has been removed');
     }
